backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncpg==0.30.0
click==8.1.8
cmdstanpy==1.2.5
contourpy==1.3.1
cycler==0.12.1
databases==0.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
fonttools==4.55.4
greenlet==3.1.1
h11==0.14.0
holidays==0.65
idna==3.10
importlib_resources==6.5.2
joblib==1.4.2
jose==1.0.0
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.1.0
prophet==1.1.6
pyasn1==0.6.1
pydantic==2.10.5
pydantic_core==2.27.2
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2024.2
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
stanio==0.5.1
starlette==0.41.3
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.34.0
